Cleanup & address comments
