+++ packages/miniflare/src/workers/core/email.ts
@@ -1,0 +1,167 @@
+import assert from "node:assert";
+import { ForwardableEmailMessage } from "@cloudflare/workers-types";
+import { blue, red, yellow } from "kleur/colors";
+import PostalMime, { Email } from "postal-mime";
+import { MiniflareEmailMessage } from "../email/email.worker";
+import { isEmailReplyable, validateReply } from "../email/validate";
+import { CoreBindings } from "./constants";
+
+type Env = {
+	[CoreBindings.SERVICE_LOOPBACK]: Fetcher;
+};
+
+function renderEmailHeaders(headers: Headers | undefined) {
+	return headers
+		? `\n  headers:\n${[...headers.entries()].map(([k, v]) => `    ${k}: ${v}`).join("\n")}`
+		: "";
+}
+
+export async function handleEmail(
+	params: URLSearchParams,
+	request: Request,
+	service: Fetcher,
+	env: Env,
+	_ctx: ExecutionContext
+): Promise<Response> {
+	// Turn an HTTP request into an EmailMessage, using:
+	//  - `from` and `to` from the URL
+	//    - These refer to the SMTP envelope addresses: https://datatracker.ietf.org/doc/html/rfc5321#section-3
+	//  - `raw` from the request body
+	//  - `headers` from the request headers
+	// Refer to https://developers.cloudflare.com/email-routing/email-workers/runtime-api/#emailmessage-definition for more details
+	const from = params.get("from");
+	const to = params.get("to");
+
+	if (!request.body || !from || !to) {
+		return new Response(
+			"Invalid email. Your request must include URL parameters specifying the `from` and `to` addresses, as well as an email in the body",
+			{
+				status: 400,
+			}
+		);
+	}
+	// We need to parse the email body in this handler in order to validate it, but we also want to pass through
+	// the raw email to the user Worker. As such, clone the request for use in this handler.
+	const clonedRequest = request.clone();
+
+	assert(clonedRequest.body !== null);
+
+	const incomingEmailRaw = new Uint8Array(await request.arrayBuffer());
+
+	// Email Routing does not support messages bigger than 25Mib: https://developers.cloudflare.com/email-routing/limits/#message-size
+	// In practice, local dev only supports 1MB, since it uses a JSRPC transport.
+	if (incomingEmailRaw.byteLength > 25 * 1024 * 1024) {
+		return new Response(
+			"Email message size is bigger than the production size limit of 25MiB. Local development has a lower limit of 1Mib.",
+			{
+				status: 400,
+			}
+		);
+	}
+	if (incomingEmailRaw.byteLength > 1024 * 1024) {
+		return new Response(
+			"Email message size is within the production size limit of 25MiB, but exceeds the lower 1Mib limit for testing locally.",
+			{
+				status: 400,
+			}
+		);
+	}
+
+	let parsedIncomingEmail: Email;
+	try {
+		parsedIncomingEmail = await PostalMime.parse(incomingEmailRaw);
+	} catch (e) {
+		const error = e as Error;
+		return new Response(
+			`Email could not be parsed: ${error.name}: ${error.message}`,
+			{ status: 400 }
+		);
+	}
+
+	if (parsedIncomingEmail.messageId === undefined) {
+		return new Response(
+			"Email could not be parsed: invalid or no message id provided",
+			{ status: 400 }
+		);
+	}
+
+	// Emails can contain both an "envelope" from/to and a "header" from/to. Warn if these are different.
+	// Refer to https://datatracker.ietf.org/doc/html/rfc5321#section-3, https://datatracker.ietf.org/doc/html/rfc5322#section-3.6.2, and https://datatracker.ietf.org/doc/html/rfc5322#section-3.6.3 for more details
+	if (from !== parsedIncomingEmail.from.address) {
+		console.warn(
+			`${yellow("Provided MAIL FROM address doesn't match the email message's \"From\" header")}:\n  MAIL FROM: ${from}\n  "From" header: ${parsedIncomingEmail.from.address}`
+		);
+	}
+
+	if (!parsedIncomingEmail.to?.map((addr) => addr.address).includes(to)) {
+		console.warn(
+			`${yellow('Provided RCPT TO address doesn\'t match any "To" header in the email message')}:\n  RCPT TO: ${to}\n  "To" header: ${parsedIncomingEmail.to?.map((addr) => addr.address).join(", ")}`
+		);
+	}
+
+	const incomingEmailHeaders = new Headers(
+		parsedIncomingEmail.headers.map((header) => [header.key, header.value])
+	);
+
+	// Propogate `.setReject()` reasons to the caller
+	let maybeClientError: string | undefined = undefined;
+
+	// @ts-expect-error .email is not in the `Fetcher` but it's a valid RPC call.
+	await service.email(
+		// Construct a ForwardableEmailMessage-like object. We need
+		// - ForwardableEmailMessage to be able to be passed across JSRPC (to support e.g. userWorker.email(ForwardableEmailMessage))
+		// - ForwardableEmailMessage properties to be synchronously available (to match production). This rules out a class extending `RpcStub`
+		// However, unlike EmailMessage (see email.worker.ts) it doesn't need to be user-constructable, and so we can just use an object with `satisfies`
+		{
+			from,
+			to,
+			raw: clonedRequest.body,
+			rawSize: incomingEmailRaw.byteLength,
+			headers: incomingEmailHeaders,
+			setReject: (reason: string): void => {
+				console.log(
+					`${red("Email handler rejected message")} with the following reason: "${reason}"`
+				);
+				maybeClientError = reason;
+			},
+			forward: async (rcptTo: string, headers?: Headers): Promise<void> => {
+				console.log(
+					`${blue("Email handler forwarded message")} with\n  rcptTo: ${rcptTo}${renderEmailHeaders(headers)}`
+				);
+			},
+			reply: async (replyMessage: MiniflareEmailMessage): Promise<void> => {
+				if (!isEmailReplyable(parsedIncomingEmail, incomingEmailHeaders)) {
+					throw new Error("Original email is not replyable");
+				}
+				const finalReply = await validateReply(
+					parsedIncomingEmail,
+					replyMessage
+				);
+
+				const resp = await env[CoreBindings.SERVICE_LOOPBACK].fetch(
+					"http://localhost/core/store-temp-file?extension=eml",
+					{
+						method: "POST",
+						body: finalReply,
+					}
+				);
+				const file = await resp.text();
+
+				console.log(
+					`${blue("Email handler replied to sender")} with the following message:\n  ${file}`
+				);
+			},
+		} satisfies ForwardableEmailMessage
+	);
+
+	if (maybeClientError !== undefined) {
+		return new Response(
+			`Worker rejected email with the following reason: ${maybeClientError}`,
+			{ status: 400 }
+		);
+	}
+
+	return new Response("Worker successfully proce", {
+		status: 200,
+	});
+}
